[WIP] working on linting
CoC sends stuff to ALE with `displayByAle`, and since nvim 0.6, ALE
sends stuff to neovim's built in Diagnostics API, meaning that many ALE
configs don't have effect, like changing the signs with
`ale_sign_error`, and maybe also filtering diagnostics (like only
showing errors and not warnings), and stuff related to virtual text (the
text of the warnings/errors etc...).

So we had to configure signs with neovim's Diagnostics API, and later,
I'll try filtering out warnings too.

Yeah, seems like quite some work to do, or at least to get the full
picture of what's going on, so I expect to try a lot of the configs in
appropriate docs of ALE and Diagnostics (links added in this commit).