Move apps installation to homebrew
It might be simpler to do everything with hombrew.
The goal of using nix-darwin/homen-manager is to be able to reproduce
my env, i.e. declaring deps in declarative way.
Moving apps installation to homebrew still enables me to do exactly
that.
And I hope that it will be simpler, since installing apps from nixospkgs
sometimes comes with problems, like missing icons in spotlight etc.